feat: Refactor supportrs to locks
